// backend/src/types.ts
export type BaseWidget = {
  id: WidgetID;
  kind: WidgetKind;
  x: number;
  y: number;
  width: number;
  height: number;
  maxHeight?: number | undefined;
};
// ...
export type ChartWidget = {
  id: WidgetID;
  kind: typeof WidgetKind.chart;
  x: number;
  y: number;
  width: number;
  height: number;
  maxHeight?: number | undefined;
  title: string;
  chartKind: ChartKind;
  direction: ChartDirection;
  segmentationKeyPaths: KeyPath[];
  series: ChartSeries[];
  segmentsTitle?: string | undefined;
};

export type ColumnDescriptor = {
  title: string;
  keyPaths: string[];
  width: number | "auto";
  description?: string | undefined;
};

export type ContentWidget = {
  id: WidgetID;
  kind: typeof WidgetKind.content;
  x: number;
  y: number;
  width: number;
  height: number;
  maxHeight?: number | undefined;
  childArtifactPath: ArtifactPath;
  showsContext: boolean;
};
// ...
export type DashboardID = UUID;

export type Dashboard = {
  id: DashboardID;
  widgets: Widget[];
};
// ...
export type MetricsListWidget = {
  id: WidgetID;
  kind: typeof WidgetKind.metricsList;
  x: number;
  y: number;
  width: number;
  height: number;
  maxHeight?: number | undefined;
  metrics: MetricDisplay[];
  groups?: string[] | undefined;
  valueColumns: MetricValuesDisplay[];
};

export type MetricsWidget = {
  id: WidgetID;
  kind: typeof WidgetKind.metrics;
  x: number;
  y: number;
  width: number;
  height: number;
  maxHeight?: number | undefined;
  metrics: MetricDisplay[];
};
// ...
export type OrganizationID = UUID;

export type Organization = {
  id: OrganizationID;
  name: string;
  template?: OrganizationTemplate | undefined;
  isDeleted?: true | undefined;
};

export const OrganizationTemplate = {
  demo: "demo",
  general: "general",
} as const;
export type OrganizationTemplate = (typeof OrganizationTemplate)[keyof typeof OrganizationTemplate];
// ...
export const TableContents = {
  artifact: "artifact",
  type: "type",
  snapshot: "snapshot",
  generation: "generation",
} as const;
export type TableContents = (typeof TableContents)[keyof typeof TableContents];

export type TableDescriptor = ColumnDescriptor[];

export type TableWidget = {
  id: WidgetID;
  kind: typeof WidgetKind.table;
  x: number;
  y: number;
  width: number;
  height: number;
  maxHeight?: number | undefined;
  contents: TableContents;
  filter?: Filter | undefined;
  columns: ColumnDescriptor[];
  showsNestedArtifacts: boolean;
};

export type TagValue = {
  tag: string;
  value: string;
};

export type URL = string;

export type UUID = string;

export type UpdateArtifactSnapshotRequest = {
  orgID: OrganizationID;
  artifactPath: ArtifactPath;
  eventSummaryID: EventSummaryID;
  snapshotDelta: ArtifactSnapshotDelta;
};

export type UpdateArtifactSnapshotResponse = {
  status: "success" | "error";
  message: string;
};

export type UserID = string;

export type User = {
  id: UserID;
  email: string;
  fullName: string;
};

export const ValueFilterOperator = {
  equal: "equal",
  notEqual: "notEqual",
  greaterThan: "greaterThan",
  lessThan: "lessThan",
  greaterThanOrEqual: "greaterThanOrEqual",
  lessThanOrEqual: "lessThanOrEqual",
} as const;
export type ValueFilterOperator = (typeof ValueFilterOperator)[keyof typeof ValueFilterOperator];

export type WidgetID = UUID;

export const WidgetKind = {
  chart: "chart",
  content: "content",
  metrics: "metrics",
  metricsList: "metricsList",
  table: "table",
} as const;
export type WidgetKind = (typeof WidgetKind)[keyof typeof WidgetKind];

export type Widget = BaseWidget & (ChartWidget | ContentWidget | MetricsWidget | MetricsListWidget | TableWidget);